+++ src/pynmeagps/nmeahelpers.py
@@ -17,6 +17,8 @@ from pynmeagps.nmeatypes_core import (
     NMEA_MSGIDS_PROP,
     LA,
     LN,
+    DT,
+    DM,
     WGS84_SMAJ_AXIS,
     WGS84_FLATTENING,
     WGS84,
@@ -179,17 +181,19 @@ def ddd2dmm(degrees: float, att: str, hpmode: bool = False) -> str:
         return ""
 
 
-def date2utc(dates: str) -> datetime.date:
+def date2utc(dates: str, form: str = DT) -> datetime.date:
     """
     Convert NMEA Date to UTC datetime.
 
-    :param str dates: NMEA date ddmmyy
+    :param str dates: NMEA date
+    :param str form: date format DT = ddmmyy, DM = mmddyy (DT)
     :return: UTC date YYyy:mm:dd
     :rtype: datetime.date
     """
 
     try:
-        utc = datetime.strptime(dates, "%d%m%y")
+        dform = "%m%d%y" if form == DM else "%d%m%y"
+        utc = datetime.strptime(dates, dform)
         return utc.date()
     except (TypeError, ValueError):
         return ""
@@ -228,17 +232,19 @@ def time2str(tim: datetime.time) -> str:
         return ""
 
 
-def date2str(dat: datetime.date) -> str:
+def date2str(dat: datetime.date, form: str = DT) -> str:
     """
     Convert datetime.date to NMEA formatted string.
 
     :param datetime.date dat: date
-    :return: NMEA formatted date string ddmmyy
+    :param str form: date format DT = ddmmyy, DM = mmddyy (DT)
+    :return: NMEA formatted date string
     :rtype: str
     """
 
     try:
-        return dat.strftime("%d%m%y")
+        dform = "%m%d%y" if form == DM else "%d%m%y"
+        return dat.strftime(dform)
     except (AttributeError, TypeError, ValueError):
         return ""